Added first actual smoke test
